Fix the Vertex AI Service to work with the latest Gemini changes
